split the engine and game

+++ Game/HealthComponent.cpp
@@ -1,4 +1,5 @@
 #include "HealthComponent.h"
+#include "GameEvent.h" 
 
 namespace dae
 {
@@ -11,11 +12,12 @@ namespace dae
         else
         {
             m_Lives -= amount;
-            m_Subject.Notify(GameEvent::PlayerDamaged, &GetOwner());
+            m_Subject.Notify(static_cast<int>(GameEvent::PlayerDamaged), &GetOwner());
         }
     }
+
     void HealthComponent::Die()
     {
-        m_Subject.Notify(GameEvent::PlayerDied, &GetOwner());
+        m_Subject.Notify(static_cast<int>(GameEvent::PlayerDied), &GetOwner());
     }
 }
